% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/marglik.R
\name{ChibML}
\alias{ChibML}
\title{Marginal Likelihood by Chib & Jeliazikov's Method for User-Written Functions}
\usage{
ChibML(logfun, theta.star, tune, V, mcmcsamp, df, verbose)
}
\arguments{
\item{logfun}{The logarithm of the objective function}

\item{theta.star}{The starting value of the inner MCMC sampling and the value required by the Chib & Jeliazikov's method. This must be a high denstiy point, such as the posterior mean, median or mode.}

\item{tune}{The tunning value to be used to achieve the desired efficiency}

\item{V}{The proposal scale matrix}

\item{mcmcsamp}{The MCMC sample from the joint posterior}

\item{df}{The degrees of freedom of the proposal}

\item{verbose}{A switch which determines whether or not the progress of the sampler is printed to the screen. If verbose is greater than 0 the iteration number, and the Metropolis acceptance rate are sent to the screen every \code{verbose}th iteration}
}
\value{
double, the logarithm of the posterior normalising constant
}
\description{
The function computes the marginal likelihood, i.e. the posterior normalising constant, with the method of Chib & Jeliazikov (2001) for user-written functions, from which an MCMC posterior sample is available.
}
\details{
The function produce an approximation of the posterior normalizing constant via the Chib & Jeliazikov method in a single block sampling. The proposal distribution for the block is a Student's \eqn{t}-density with \code{df} degrees of freedom. The proposal is centered at the current value of \eqn{\theta}{theta} and has scale matrix \eqn{H}. \eqn{H} is calculated as: \eqn{H = TVT}{H = T*V*T}, where \eqn{T}{T} is a the diagonal positive definite matrix formed from the \code{tune}.
}
\references{
Chib S. & Jeliazikov I. (2001).
Marginal likelihood from the Metropolis-Hastings output.
\emph{Journal of the American Statistical Association}, \bold{46},
 270--281.

Robert C. P. & Casella G. (2004).
 \emph{Monte Carlo Statistical Methods}. 2nd Edition. New York: Springer.
}
\seealso{
\code{\link[iLaplaceExamples]{nlpost_gomp}} and \code{\link[iLaplaceExamples]{nlpost_bod2}} for examples; \code{\link[iLaplaceExamples]{MHmcmc}}, \code{\link[iLaplaceExamples]{isML}}
}

