% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iLaplaceExtra.R
\docType{package}
\name{iLaplaceExtra}
\alias{iLaplaceExtra}
\alias{iLaplaceExtra-package}
\title{iLaplaceExtra: A package for computating the notorious bar statistic.}
\description{
The iLaplaceExtra package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions...
}

