% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iLaplaceExamples.R
\docType{package}
\name{iLaplaceExtra}
\alias{iLaplaceExtra}
\alias{iLaplaceExtra-package}
\title{iLaplaceExamples: A package for examples with iLaplace.}
\description{
The iLaplaceExamples package provides utilites for running examples of approximating marignal likelihoods by the improved Laplace method of Ruli et al. (2015), implemented in the package iLaplace. Other approximation method such as the standard Laplace and some Monte Carlo methods are aslo provided. There are five examples:
\enumerate{
 \item Gompertz model
 \item multivariate t/skew-t distribtuion
 \item nonlinear regression models with two datasets
 \item binary logistic regression with random effects
 \item binary logistic regression with crossed random effects with the \code{\link[iLaplaceExamples]{Salamander}} data}
which functions are listed below.
foo, bar and baz.
}
\section{Functions for Gomper}{

The foo functions...
}

